fix: resolve TS errors after merge from main
