+++ web_src/.eslint-budget-baseline.json
@@ -1,19 +1,21 @@
 {
-  "maxAllowedTotalIssues": 657,
+  "maxAllowedTotalIssues": 666,
   "maxAllowedByRule": {
-    "complexity": 232,
+    "complexity": 233,
     "@typescript-eslint/no-explicit-any": 160,
     "@typescript-eslint/no-non-null-asserted-optional-chain": 80,
-    "max-lines-per-function": 76,
+    "max-lines-per-function": 79,
     "max-statements": 74,
-    "max-lines": 17,
+    "max-lines": 18,
     "max-params": 6,
     "max-depth": 6,
-    "no-eslint-disable-next-line": 6
+    "no-eslint-disable-next-line": 6,
+    "react-hooks/exhaustive-deps": 2,
+    "react/jsx-no-bind": 2
   },
   "maxAllowedMetricMaximumByRule": {
-    "max-lines": 5404,
-    "complexity": 340
+    "max-lines": 5239,
+    "complexity": 342
   },
-  "updatedAt": "2026-05-17T21:43:50.665Z"
+  "updatedAt": "2026-05-18T07:34:03.025Z"
 }
