+++ web_src/src/components/AgentSidebar/widgets/RubricWidget.tsx
@@ -36,9 +36,7 @@ export function RubricWidget({ title, criteria, onAction, onStartBuilding }: Rub
         <div className="px-3 py-2 bg-violet-50 border-b border-violet-200 flex items-center gap-2">
           <ClipboardList size={14} className="text-violet-600 shrink-0" />
           <p className="text-xs font-semibold text-violet-900 flex-1">{title || "Build Plan"}</p>
-          <span className="text-[10px] text-violet-500 font-medium">
-            {criteria.length} criteria
-          </span>
+          <span className="text-[10px] text-violet-500 font-medium">{criteria.length} criteria</span>
         </div>
 
         {/* Criteria list */}
@@ -55,8 +53,7 @@ export function RubricWidget({ title, criteria, onAction, onStartBuilding }: Rub
               onClick={() => setExpanded(true)}
               className="flex items-center gap-1 text-[10px] text-violet-500 hover:text-violet-700 mt-1"
             >
-              <ChevronDown size={10} />
-              +{criteria.length - PREVIEW_COUNT} more
+              <ChevronDown size={10} />+{criteria.length - PREVIEW_COUNT} more
             </button>
           )}
           {hasMore && expanded && (
@@ -73,12 +70,7 @@ export function RubricWidget({ title, criteria, onAction, onStartBuilding }: Rub
 
         {/* Actions */}
         <div className="px-3 pb-3 pt-1 flex items-center gap-2 border-t border-violet-100">
-          <Button
-            variant="ghost"
-            size="sm"
-            className="text-xs text-slate-500 h-7"
-            onClick={() => setModalOpen(true)}
-          >
+          <Button variant="ghost" size="sm" className="text-xs text-slate-500 h-7" onClick={() => setModalOpen(true)}>
             View Full Plan
           </Button>
           <Button
@@ -100,11 +92,7 @@ export function RubricWidget({ title, criteria, onAction, onStartBuilding }: Rub
                 <ClipboardList size={16} className="text-violet-600" />
                 <h2 className="text-sm font-semibold text-slate-900">{title || "Build Plan"}</h2>
               </div>
-              <button
-                type="button"
-                onClick={() => setModalOpen(false)}
-                className="text-slate-400 hover:text-slate-600"
-              >
+              <button type="button" onClick={() => setModalOpen(false)} className="text-slate-400 hover:text-slate-600">
                 <X size={16} />
               </button>
             </div>
@@ -117,12 +105,7 @@ export function RubricWidget({ title, criteria, onAction, onStartBuilding }: Rub
               ))}
             </div>
             <div className="px-4 py-3 border-t border-slate-200 flex justify-end">
-              <Button
-                variant="ghost"
-                size="sm"
-                className="text-xs"
-                onClick={() => setModalOpen(false)}
-              >
+              <Button variant="ghost" size="sm" className="text-xs" onClick={() => setModalOpen(false)}>
                 Close
               </Button>
             </div>
